Update Header to be NavBar. Move CustomIcon to components
